se agrega funcionalidades a las cuentas, para transferir, depositar y retirar, se registran logs
